requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.39.0
Pillow>=10.0.0
numpy>=1.24.0
runpod>=1.6.0
